RavenDB-25738 Group backward-compatibility escape/ID helpers under Compatibility
