+++ src/Raven.Server/Documents/DocumentPutAction.cs
@@ -126,7 +126,7 @@ namespace Raven.Server.Documents
             }
 
             id = BuildDocumentId(id, newEtag, out bool knownNewId);
-            using (DocumentIdWorker.GetLowerIdSliceAndStorageKeyForBackwardCompatibility(context, id, out Slice lowerId, out Slice idPtr))
+            using (DocumentIdWorker.Compatibility.GetLowerIdSliceAndStorageKey(context, id, out Slice lowerId, out Slice idPtr))
             {
                 if (newFlags.HasFlag(DocumentFlags.FromResharding) == false)
                     _documentsStorage.ValidateId(context, lowerId, type: DocumentChangeTypes.Put, newFlags);
