+++ src/Sparrow/Json/JsonOperationContext.cs
@@ -1217,32 +1217,6 @@ namespace Sparrow.Json
             return allocatedArray;
         }
 
-#if NET8_0_OR_GREATER
-        public unsafe LazyStringValue GetLazyStringForBackwardCompatibility(byte* ptr, int size, bool longLived = false)
-        {
-            var state = new JsonParserState();
-            var maxByteCount = Encodings.Utf8.GetMaxByteCount(size);
-
-            int escapePositionsSize = StringUtils.FindMaxEscapePositionAndControlCharSizeForBackwardCompatibility(ptr, size, out _);
-
-            int memorySize = maxByteCount + escapePositionsSize;
-            var memory = longLived ? GetLongLivedMemory(memorySize) : GetMemory(memorySize);
-
-            var address = memory.Address;
-
-            Memory.Copy(address, ptr, size);
-
-            state.FindEscapedPositionsAndEscapeControlsForBackwardCompatibility(address, ref size, escapePositionsSize);
-
-            state.WriteEscapePositionsTo(address + size);
-            LazyStringValue result = longLived == false ? AllocateStringValue(null, address, size) : new LazyStringValue(null, address, size, this);
-            result.AllocatedMemoryData = memory;
-
-            result.EscapePositions = state.EscapePositions.Count > 0 ? state.EscapePositions.ToArray() : [];
-            return result;
-        }
-#endif
-        
 #if DEBUG || VALIDATE
 
         private sealed class IntReference
