+++ src/Sparrow/Json/Parsing/JsonParserState.cs
@@ -72,13 +72,5 @@ namespace Sparrow.Json.Parsing
             EscapePositions.Clear();
         }
         
-#if NET8_0_OR_GREATER
-        //RavenDB-25738 For backward compatibility only. Use only for document IDs, collection, attachment name & type, and timeseries tag
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public void FindEscapedPositionsAndEscapeControlsForBackwardCompatibility(byte* str, ref int len, int previousComputedMaxSize)
-        {
-            StringUtils.FindEscapedPositionsAndEscapeControlsForBackwardCompatibility(EscapePositions, str, ref len, previousComputedMaxSize);
-        }
-#endif
     }
 }
