+++ test/SlowTests/Issues/RavenDB-24797.cs
@@ -662,7 +662,7 @@ namespace SlowTests.Issues
                     using (var clonedKey = context.AllocateStringValue(null, buffer.Ptr, buffer.Length))
                     using (Slice.External(context.Allocator, clonedKey, out var countersGroupKey))
                     using (Slice.From(context.Allocator, changeVector, out var cv))
-                    using (DocumentIdWorker.GetLowerIdSliceAndStorageKeyForBackwardCompatibility(context, "Users", out _, out Slice collectionSlice))
+                    using (DocumentIdWorker.Compatibility.GetLowerIdSliceAndStorageKey(context, "Users", out _, out Slice collectionSlice))
                     using (database.DocumentsStorage.ContextPool.AllocateOperationContext(out DocumentsOperationContext writeContext))
                     using (var tx = writeContext.OpenWriteTransaction())
                     {
